+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [2.2.0] - 13/04/2021
+ - Added [backTextStyle] to change the text style for the back button
+  - Added [nextTextStyle] to change the text style for the next button
+
 ## [2.1.0] - 13/04/2021
  - Deprecated [iconColor], use [icon] instead.
  It will be removed in the next major version.